+++ static/javascript/content/css.js
@@ -9,6 +9,12 @@ define(['Backbone'], function (Backbone) {
 	return Backbone.Collection.extend({
 		content: [
 			{
+				title: 'CSS翻转动画',
+				date: '2016-06-02',
+				url: '../static/notes/css/animation-reverse.html',
+				des: 'CSS翻转动画'
+			},
+			{
 				title: '页面布局',
 				date: '2016-01-19',
 				url: '../static/notes/css/layout.html',