+++ webpackVersion/scripts/content/notes.js
@@ -1,4 +1,4 @@
-const nodejs = {
+module.exports = {
 	title: 'Nodejs',
 	contents: [
 		{
@@ -18,5 +18,3 @@ const nodejs = {
 		}
 	]
 };
-
-export default nodejs;